nd4j version -> stable
Former-commit-id: 9655daba88b78027dd7d6719674ed2b1ab66d0a8 [formerly b7a88502f9ac8a31288e33b610aed19a22074ae4]
Former-commit-id: b37993591a9317b6daa7b626e5ebe1b561a2e148

+++ deeplearning4j-scaleout/deeplearning4j-scaleout-akka/src/main/java/org/deeplearning4j/scaleout/testsupport/BaseTestDistributed.java
@@ -39,7 +39,8 @@ public abstract class BaseTestDistributed {
         stateTracker.finish();
         stateTracker.shutdown();
         distributed.shutdown();
-        new File("model-saver").delete();
+        if(new File("model-saver").exists())
+            new File("model-saver").delete();
         Thread.sleep(10000);
     }
